fix(navigation-menu): simplify viewport content existence validation
Sometimes the `DynamicDelegateRenderer` will mark an element for removal without actually removing the element. This will cause the check for a child element (in this case, `div.NavigationMenuContentWrapper`) to always be true, and cause an infinite loop.

This adjusts the `updateActiveContent` function to only check for existing viewport content once to avoid these eventualities.